hwmon: (corsair-psu) update supported devices
Adds support for another Corsair PSUs series: AX760i, AX860i, AX1200i,
AX1500i and AX1600i. The first 3 power supplies are supported through
the Corsair Link USB Dongle which is some kind of USB/Serial/TTL
converter especially made for the COM ports of these power supplies.
There are 3 known revisions of these adapters. The AX1500i power supply
has revision 3 built into the case and AX1600i is the only one in that
series, which has an unique usb hid id like the RM/RX series.

The patch also changes the usb hid ids to use upper case letters to be
consistent with the rest of the hex numbers in the driver and updates
the hwmon documentation.

This patch adds:
- hwmon/corsair-psu documentation update
- corsair-psu driver update

Signed-off-by: Wilken Gottwalt <wilken.gottwalt@posteo.net>
Link: https://lore.kernel.org/r/X7+T4aZSUuzfsf7H@monster.powergraphx.local
Signed-off-by: Guenter Roeck <linux@roeck-us.net>